requirements.txt
# --- 核心算法 (YOLOv11 & Torch) ---
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0

# --- 数据处理与抽帧工具 ---
opencv-python>=4.8.0
numpy>=1.23.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0

# --- OCR 识别 (PaddleOCR) ---
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# --- 后端框架 (FastAPI) ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart==0.0.9
pydantic>=2.0.0
python-dotenv>=1.0.0

# --- 通用工具 ---
requests>=2.31.0
httpx>=0.27.0
pillow>=10.0.0
pyyaml>=6.0.1

# --- 增强推理引擎 (SAHI 切片推理) ---
sahi>=0.11.15

# --- 数据库核心 ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# --- 密码加密与 JWT 鉴权 ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2

# --- 聚类算法 ---
scikit-learn>=1.3.0